refactor(config): move ConfigMigrator from domain to infrastructure layer
ConfigMigrator handles migration of legacy configuration formats,
which is a technical/infrastructure concern rather than a business rule.

Changes:
- Create infrastructure/config/migration.py with full ConfigMigrator
- Update domain/configs/migration.py to deprecation stub with re-export
- Update domain/configs/__init__.py with lazy loading and DeprecationWarning
- Update domain/configs/pipeline.py to use lazy import from infrastructure
- Update infrastructure/config/loader.py to import from new location
- Export ConfigMigrator from infrastructure/config/__init__.py

This follows DDD principles by keeping domain layer focused on
business rules while infrastructure handles technical concerns
like format versioning and migration.

All existing tests pass. Backward compatibility maintained via
deprecation stubs that emit DeprecationWarning.
