+++ src/bioetl/domain/configs/pipeline.py
@@ -673,7 +673,10 @@ class PipelineConfig(BaseModel):
     @classmethod
     def migrate_legacy_format(cls, data: dict[str, Any]) -> dict[str, Any]:
         """Delegate migration to ConfigMigrator."""
-        from bioetl.domain.configs.migration import ConfigMigrator
+        # Lazy import from infrastructure layer to avoid circular imports
+        # ConfigMigrator was moved from domain to infrastructure as migration
+        # is a technical concern, not a business rule
+        from bioetl.infrastructure.config.migration import ConfigMigrator
 
         return ConfigMigrator.migrate(data)
